Move AppRepository up from package db and add Tiro network API controllers
